requirements.txt
Jinja2
junitxml
fasteners
netaddr
paramiko
pexpect
PyYAML
pymongo
python-keystoneclient
testtools==1.7.1
xunitparser
python-fuelclient
python-muranoclient
oslo.utils<1.5.0
oslo.config>=1.2.1,<=1.11.0
oslo.serialization>=1.0.0,<1.5.0
oslo.i18n>=1.3.0,<1.6.0
ospurge
testrepository>=0.0.18
sqlalchemy
psycopg2
rally==0.4.0
influxdb==2.12.0
python-subunit
extras
configparser==3.5.0
scp
